main: move cert tests to a separated file
